fip: move headers shared between TF and fiptool to include/host_api
Some header files need to be shared between TF and host programs.
For fiptool, two headers are copied to the tools/fiptool directory,
but it looks clumsy.

This commit introduces a new directory include/host_api, which
collects headers that should be shared between TF and host programs.

This will clarify the interface exposed to host tools.  We should
add new headers to this directory only when we really need to do so.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>