cert: move platform_oid.h to include/host_api for all platforms
platform_oid.h is supposed to define each vendor's own OIDs, but
it does not add any value, but just burdens; vendors need to obtain
their own OIDs if they do not have ones, and duplicate platform_oid.h
to define the same macros with different OIDs.

All vendors should be able to share the same header (i.e. OIDs of
ARM Ltd.) for this purpose because they just re-use the extention
fields defined by ARM Ltd.

The platform_oid.h has been renamed to tbbr_oid.h because it is
no longer platform-dependent, and moved to include/host_api so that
it can be shared between TF and the cert_create tool.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>